src/backend/requirements.txt

fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
psycopg2-binary==2.9.5
pydantic==1.10.7
python-dotenv==1.0.0
azure-identity==1.13.0
openai==1.12.0
python-multipart==0.0.6
jose==1.0.0
passlib==1.7.4
bcrypt==4.0.1
alembic==1.10.2
